chore: fixes and improvemants
