+++ apps/asana/.env.test
@@ -7,4 +7,4 @@ ASANA_API_BASE_URL="https://app.asana.com/api/1.0"
 ASANA_USERS_SYNC_CRON="0 0 * * *"
 NANGO_INTEGRATION_ID="asana-sandbox"
 NANGO_SECRET_KEY="nango-secret-key"
-date-fns="elba-webhook-secret-docusign"
+
