+++ apps/asana/src/connectors/asana/users.test.ts
@@ -4,12 +4,12 @@ import { server } from '@elba-security/test-utils';
 import { env } from '@/common/env';
 import { AsanaError } from '../common/error';
 import type { AsanaUser } from './users';
-import { getUsers, deleteUser, getAuthUser, getWorkspaceIds } from './users';
+import { getUsers, deleteUser, getAuthUser } from './users';
 
 const validToken = 'token-1234';
 const endPage = '3';
 const nextPage = '2';
-const userIds = 'test-user-id';
+const userId = 'test-user-id';
 const workspaceId = '000000';
 const invalidToken = 'invalid-token';
 const authUserId = 'test-auth-user-id';
@@ -26,7 +26,6 @@ const invalidUsers = [];
 
 describe('users connector', () => {
   describe('getUsers', () => {
-    // mock token API endpoint using msw
     beforeEach(() => {
       server.use(
         http.get(`${env.ASANA_API_BASE_URL}/users`, ({ request }) => {
@@ -83,48 +82,20 @@ describe('users connector', () => {
 
     test('should delete user successfully when token is valid', async () => {
       await expect(
-        deleteUser({ accessToken: validToken, workspaceId, userIds })
+        deleteUser({ accessToken: validToken, workspaceId, userId })
       ).resolves.not.toThrow();
     });
 
     test('should not throw when the user is not found', async () => {
       await expect(
-        deleteUser({ accessToken: validToken, workspaceId, userIds })
+        deleteUser({ accessToken: validToken, workspaceId, userId })
       ).resolves.toBeUndefined();
     });
 
     test('should throw AsanaError when token is invalid', async () => {
       await expect(
-        deleteUser({ accessToken: 'invalidToken', workspaceId, userIds })
+        deleteUser({ accessToken: 'invalidToken', workspaceId, userId })
       ).rejects.toBeInstanceOf(AsanaError);
-    });
-  });
-
-  describe('getWorkspaceIds', () => {
-    beforeEach(() => {
-      server.use(
-        http.get(`${env.ASANA_API_BASE_URL}/workspaces`, ({ request }) => {
-          if (request.headers.get('Authorization') !== `Bearer ${validToken}`) {
-            return new Response(undefined, { status: 401 });
-          }
-
-          return Response.json({
-            data: [
-              {
-                gid: workspaceId,
-              },
-            ],
-          });
-        })
-      );
-    });
-
-    test('should return the workspaceIds when the accessToken is valid', async () => {
-      await expect(getWorkspaceIds(validToken)).resolves.toStrictEqual([workspaceId]);
-    });
-
-    test('should throw when the accessToken is invalid', async () => {
-      await expect(getWorkspaceIds(invalidToken)).rejects.toBeInstanceOf(AsanaError);
     });
   });
