+++ apps/asana/src/connectors/asana/users.ts
@@ -22,14 +22,6 @@ const asanaResponseSchema = z.object({
     .optional(),
 });
 
-const workspaceResponseSchema = z.object({
-  data: z.array(
-    z.object({
-      gid: z.string(),
-    })
-  ),
-});
-
 const getAuthUserResponseData = z.object({
   data: z.object({
     gid: z.string(),
@@ -94,7 +86,7 @@ export const getUsers = async ({ accessToken, page }: GetUsersParams) => {
   };
 };
 
-export const deleteUser = async ({ userId, workspaceId, accessToken }: DeleteUsersParams) => {
+export const deleteUser = async ({ accessToken, workspaceId, userId }: DeleteUsersParams) => {
   const response = await fetch(`${env.ASANA_API_BASE_URL}/workspaces/${workspaceId}/removeUser`, {
     method: 'post',
     headers: {
@@ -117,40 +109,6 @@ export const deleteUser = async ({ userId, workspaceId, accessToken }: DeleteUse
   }
 };
 
-export const getWorkspaceIds = async (accessToken: string) => {
-  const response = await fetch(`${env.ASANA_API_BASE_URL}/workspaces`, {
-    method: 'get',
-    headers: {
-      Authorization: `Bearer ${accessToken}`,
-      'Content-Type': 'application/json',
-    },
-  });
-
-  if (!response.ok) {
-    throw new AsanaError('Could not retrieve workspace', { response });
-  }
-
-  const resData: unknown = await response.json();
-
-  const result = workspaceResponseSchema.safeParse(resData);
-
-  if (!result.success) {
-    throw new AsanaError('Could not parse workspace response');
-  }
-
-  if (result.data.data.length === 0) {
-    throw new AsanaError('No workspace found');
-  }
-
-  const workspaceIds = result.data.data.map((board) => board.gid);
-
-  if (!workspaceIds.length) {
-    throw new AsanaError('No Main workspace found');
-  }
-
-  return workspaceIds;
-};
-
 export const getAuthUser = async (accessToken: string) => {
   const url = new URL(`${env.ASANA_API_BASE_URL}/users/me`);
 
@@ -169,6 +127,7 @@ export const getAuthUser = async (accessToken: string) => {
   const resData: unknown = await response.json();
 
   const result = getAuthUserResponseData.safeParse(resData);
+
   if (!result.success) {
     logger.error('Invalid Asana auth-user id response', { resData });
     throw new AsanaError('Invalid Asana auth-user id response');
